Refactor manual API test script into modules
